Allocate buffers dynamically as needed instead of as part of fdinfo struct
Previously we'd allocate a buffer of 32k each for each fd up to the
compiled in StaticMaxFD, which could be as large as 65535 on some
systems (that's the Solaris 64-bit default FD_SETSIZE), resulting in
gigabytes of bss allocation on startup.

Now we don't allocate until we're actually setting up the fdinfo for
use as part of a client<->server connection.
Reviewed-by: Jeremy Huddleston <jeremyhu@apple.com>