Bug 46689 - xscope fd_set debug output is undefined
https://bugs.freedesktop.org/show_bug.cgi?id=46689

Change fd_set debug output to display actual contents of first 32 or 64 bits.

In xscope-1.3.1, MainLoop() contains debug(128,...) code that appears
intended to display the fd_set contents before and after the main
select() call.  In most implementations, fd_set will be a struct though
it is printed as an int, thus the actual debug output is unlikely to
usefully reflect the actual fd_set contents.

Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>