Bug 46696 - Try alternate protocol on X11 connection failure.
In USE_XTRANS mode, MakeConnection() cycles through a selection of
protocol choices until it finds one where it can open a socket to
the server.  It then attempts an X11 connection via that socket and
aborts if one isn't possible.

This patch modifies the code to try an alternate protocol if the X11
connection fails.

https://bugs.freedesktop.org/show_bug.cgi?id=46696

Reviewed-by: Alan Coopersmith <alan.coopersmith@oracle.com>
Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>